Move FakeBitmapPool out of sharedTest.
