+++ coil-base/src/sharedTest/java/coil/util/SharedTestFunctions.kt
@@ -24,6 +24,7 @@ import okhttp3.Headers
 import okhttp3.OkHttpClient
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
+import okio.Buffer
 import okio.buffer
 import okio.source
 
@@ -50,7 +51,8 @@ internal fun createImageLoader(
 fun createMockWebServer(context: Context, vararg images: String): MockWebServer {
     return MockWebServer().apply {
         images.forEach { image ->
-            val buffer = context.assets.open(image).source().buffer().buffer
+            val buffer = Buffer()
+            context.assets.open(image).source().buffer().readAll(buffer)
             enqueue(MockResponse().setBody(buffer))
         }
         start()
